+++ R/libs.R
@@ -1,7 +1,7 @@
 #' Load multiple packages and install if needed
 #'
 #' Install and load packages "on the fly".
-#' 
+#'
 #' The 'agridat' package uses dozens of packages in the examples for each dataset.
 #' The 'libs' function provides a simple way to load multiple packages at once,
 #' and can install any missing packages on-the-fly.
@@ -11,22 +11,24 @@
 #' @param ... Comma-separated unquoted package names
 #' @return None
 #' @author Kevin Wright
-#' @examples 
+#' @examples
 #' \dontrun{
 #' libs(dplyr,reshape2)
 #' }
-#' @references 
+#' @references
 #' None
 #' @importFrom utils install.packages installed.packages
-#' @export 
-libs <- function (...) {
+#' @export
+libs <- function(...) {
   args <- as.list(match.call())[-1]
-  packs <- sapply(args, function(x)
-    if (is.character(x)) x else deparse(x))
+  packs <- sapply(args, function(x) {
+    if (is.character(x)) x else deparse(x)
+  })
   ip <- rownames(installed.packages())
   for (p in packs) {
-    if (!(p %in% ip)) 
+    if (!(p %in% ip)) {
       install.packages(p)
+    }
     library(p, character.only = TRUE)
   }
 }